# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AXYNTRAX Dashboard Supremo — API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments:
    get:
      operationId: listPayments
      tags: [finance]
      summary: Lista pagos
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: clientId
          schema: { type: integer }
      responses:
        "200":
          description: Pagos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [finance]
      summary: Registra un pago manual
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentCreate"
      responses:
        "201":
          description: Pago creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/culqi/charge:
    post:
      operationId: culqiCharge
      tags: [finance]
      summary: Genera un cargo en Culqi (modo stub si CULQI_SECRET_KEY no existe)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CulqiChargeRequest"
      responses:
        "201":
          description: Cargo creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}/sunat-comprobante:
    post:
      operationId: generateSunatComprobante
      tags: [finance]
      summary: Genera el XML UBL 2.1 (boleta/factura) del pago
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SunatComprobanteRequest"
      responses:
        "200":
          description: Comprobante generado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SunatComprobanteResponse"
...
  /push/subscribe:
    post:
      operationId: subscribePush
      summary: Registra una suscripción Web Push
      tags: [push]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscribeRequest"
      responses:
        "200":
          description: Suscripción registrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  id: { type: integer }
                  updated: { type: boolean }
                required: [ok, id, updated]
...
components:
  schemas:
...
    Payment:
      type: object
      properties:
        id: { type: integer }
        clientId: { type: ["integer", "null"] }
        licenseId: { type: ["integer", "null"] }
        amount: { type: number }
        currency: { type: string }
        method: { type: string }
        status: { type: string, enum: [pendiente, procesando, exitoso, fallido, reembolsado] }
        reference: { type: ["string", "null"] }
        externalId: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        paidAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
      required: [id, amount, currency, method, status, createdAt]
    PaymentCreate:
      type: object
      properties:
        clientId: { type: ["integer", "null"] }
        licenseId: { type: ["integer", "null"] }
        amount: { type: number }
        currency: { type: ["string", "null"] }
        method: { type: ["string", "null"] }
        status: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
      required: [amount]
    CulqiChargeRequest:
      type: object
      properties:
        amount: { type: number }
        currency: { type: ["string", "null"] }
        clientId: { type: ["integer", "null"] }
        licenseId: { type: ["integer", "null"] }
        description: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        culqiToken: { type: ["string", "null"] }
      required: [amount]
    SunatComprobanteRequest:
      type: object
      properties:
        tipo: { type: string, enum: [boleta, factura] }
        rucCliente: { type: ["string", "null"] }
        razonSocial: { type: ["string", "null"] }
      required: [tipo]
    SunatComprobanteResponse:
      type: object
      properties:
        serie: { type: string }
        correlativo: { type: integer }
        tipo: { type: string, enum: [boleta, factura] }
        xml: { type: string }
        sentToSunat: { type: boolean }
        sunatNote: { type: ["string", "null"] }
      required: [serie, correlativo, tipo, xml, sentToSunat]
...
    PushSubscribeRequest:
      type: object
      properties:
        endpoint: { type: string }
        keys:
          type: object
          properties:
            p256dh: { type: string }
            auth: { type: string }
          required: [p256dh, auth]
      required: [endpoint, keys]
...