# lib/api-zod/src/generated/types/culqiChargeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AXYNTRAX Dashboard Supremo — API
 * OpenAPI spec version: 0.1.0
 */

export interface CulqiChargeRequest {
  amount: number;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  licenseId?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  culqiToken?: string | null;
}