# lib/api-zod/src/generated/types/sunatComprobanteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AXYNTRAX Dashboard Supremo — API
 * OpenAPI spec version: 0.1.0
 */
import type { SunatComprobanteRequestTipo } from "./sunatComprobanteRequestTipo";

export interface SunatComprobanteRequest {
  tipo: SunatComprobanteRequestTipo;
  /** @nullable */
  rucCliente?: string | null;
  /** @nullable */
  razonSocial?: string | null;
}